requirements.txt
fastapi
uvicorn[standard]
httpx
openai
pydantic
pyyaml
geopy
